# docs/design/acm-v1.0a-openapi-001.yaml
...
openapi: 3.1.0
info:
  title: ACM v1.0a — Academy Custom Module API
  version: 1.0.0
  description: |
    Trinity Academy Custom Module on AMB platform.
    Modules: CSL (신규상담) · DSH (대시보드) · SCH (학교) · REF (참조자료) · QNA (정기상담).
  contact:
    name: 김태윤 팀장
  license:
    name: Proprietary
servers:
  - url: https://api.amb.example.com/api/acm
    description: Production
  - url: https://stg-api.amb.example.com/api/acm
    description: Staging
  - url: http://localhost:3000/api/acm
    description: Local
...
security:
  - bearerAuth: []

# ─────────────────────────────────────────────
# Paths
# ─────────────────────────────────────────────
paths:
...
  /sch/schools/autocomplete:
    get:
      tags: [sch-schools]
      operationId: autocompleteSchools
      summary: Autocomplete (S-10) — pg_trgm cached
      parameters:
        - { in: query, name: q, required: true, schema: { type: string, minLength: 1 } }
        - { in: query, name: limit, schema: { type: integer, default: 10, maximum: 20 } }
      responses: { '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/SchoolSummary' } } } } } }
  /sch/schools/{schId}:
    parameters: [{ in: path, name: schId, required: true, schema: { type: string, format: uuid } }]
    get: { tags: [sch-schools], operationId: getSchool, summary: Detail (S-03), responses: { '200': { description: OK } } }
    patch: { tags: [sch-schools], operationId: updateSchool, summary: Update (S-04), responses: { '200': { description: OK } } }
    delete: { tags: [sch-schools], operationId: deleteSchool, summary: Soft-delete (S-05), responses: { '204': { description: Deleted }, '422': { description: Refs exist } } }
...
  /qna/records/{qnaId}/promote-faq:
    parameters: [{ $ref: '#/components/parameters/QnaId' }]
    post:
      tags: [qna-faq]
      operationId: promoteFaq
      summary: Promote to FAQ (Q-21, team_lead+)
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PromoteFaqDto' } } }
      responses: { '200': { description: OK }, '422': { description: BIZ_FAQ_NOT_RESOLVED } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    QnaId:
      in: path
      name: qnaId
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    SchoolSummary:
      type: object
      properties:
        schId: { type: string, format: uuid }
        schName: { type: string }
        schType: { type: string }
        schRegion: { type: string }
...
    PromoteFaqDto:
      type: object
      required: [qnaFaqVisibility]
      properties:
        qnaFaqVisibility:
          type: string
          enum: [ADVISOR_ONLY, ALL_USER, INCLUDE_TEACHER]
...